Changed u_long to unsigned long for gcc 4.0.  Changed 0 to NULL in variable-argument X functions to avoid 64-bit problems (as contributed by Ernest Williams).

+++ StripMisc.c
@@ -109,7 +109,7 @@ char    *time2str (struct timeval *t)
 {
   static char buf[512];
 
-  sprintf (buf, "sec:\t %lu\nusec:\t %ld", t->tv_sec, t->tv_usec);
+  sprintf (buf, "sec:\t %lu\nusec:\t %ld", t->tv_sec, (long)t->tv_usec);
   return buf;
 }
